+++ R/features_method.R
@@ -32,11 +32,30 @@ setGeneric("features", function(x, y, ...) standardGeneric("features"))
 #' Manning, Christopher D.; Schuetze, Hinrich (1999): \emph{Foundations of Statistical Natural Language
 #' Processing}. MIT Press: Cambridge, Mass., pp. 151-189 (ch. 5).
 #' @examples
-#'   use("polmineR")
-#'   kauder <- partition("GERMAPARLMINI", speaker = "Volker Kauder", pAttribute="word")
-#'   all <- partition("GERMAPARLMINI", date = ".*", regex = TRUE, pAttribute = "word")
-#'   terms_kauder <- features(x = kauder, y = all, included = TRUE)
-#'   top100 <- subset(terms_kauder, rank_chisquare <= 100)
+#' use("polmineR")
+#' 
+#' kauder <- partition(
+#'   "GERMAPARLMINI",
+#'   speaker = "Volker Kauder", interjection = "speech",
+#'   pAttribute="word"
+#'   )
+#' all <- partition("GERMAPARLMINI", interjection = "speech", pAttribute = "word")
+#'
+#' terms_kauder <- features(x = kauder, y = all, included = TRUE)
+#' top100 <- subset(terms_kauder, rank_chisquare <= 100)
+#' head(top100)
+#' 
+#' # a different way is to compare count objects
+#' kauder_count <- as(kauder, "count")
+#' all_count <- as(all, "count")
+#' terms_kauder <- features(kauder_count, all_count, included = TRUE)
+#' top100 <- subset(z, rank_chisquare <= 100)
+#' head(top100)
+#' 
+#' speakers <- partitionBundle("GERMAPARLMINI", sAttribute = "speaker")
+#' speakers <- enrich(speakers, pAttribute = "word")
+#' speaker_terms <- features(speakers[[1:5]], all, included = TRUE, progress = TRUE)
+#' dtm <- as.DocumentTermMatrix(speaker_terms, col = "chisquare")
 #' @rdname  features-method
 setMethod("features", "partition", function(
   x, y,
@@ -73,14 +92,6 @@ setMethod("features", "partition", function(
 })
 
 
-#' @examples
-#'   use("polmineR")
-#'   kauder <- partition("GERMAPARLMINI", speaker = "Volker Kauder", pAttribute = "word")
-#'   x <- as(kauder, "count")
-#'   all <- partition("GERMAPARLMINI", date = ".*", regex = TRUE, pAttribute = "word")
-#'   y <- as(all, "count")
-#'   z <- features(x, y, included = TRUE)
-#'   top100 <- subset(z, rank_chisquare <= 100)
 #' @rdname  features-method
 setMethod("features", "count", function(x, y, by = NULL, included = FALSE, method = "chisquare", verbose = TRUE){
   stopifnot(
@@ -113,15 +124,7 @@ setMethod("features", "count", function(x, y, by = NULL, included = FALSE, metho
 })
 
 
-#' @docType methods
 #' @rdname features-method
-#' @examples 
-#'   use("polmineR")
-#'   byName <- partitionBundle("GERMAPARLMINI", sAttribute = "speaker")
-#'   byName <- enrich(byName, pAttribute = "word")
-#'   all <- partition("GERMAPARLMINI", date = ".*", regex = TRUE, pAttribute = "word")
-#'   result <- features(byName, all, included = TRUE, progress = TRUE)
-#'   dtm <- as.DocumentTermMatrix(result, col = "chisquare")
 setMethod("features", "partitionBundle", function(
   x, y, 
   included = FALSE, method = "chisquare", verbose = TRUE, mc = getOption("polmineR.mc"), progress = FALSE